+++ src/main/java/com/itbangmodkradankanbanapi/controller/TaskController.java
@@ -1,9 +1,10 @@
-package com.itbangmodkradankanbanapi.v1.controller;
+package com.itbangmodkradankanbanapi.controller;
 
-import com.itbangmodkradankanbanapi.v1.dto.TaskDTO;
-import com.itbangmodkradankanbanapi.v1.entities.Task;
+import com.itbangmodkradankanbanapi.dto.TaskDTO;
+import com.itbangmodkradankanbanapi.dto.TaskDTOForAdd;
+import com.itbangmodkradankanbanapi.entities.Task;
 import com.itbangmodkradankanbanapi.service.ListMapper;
-import com.itbangmodkradankanbanapi.v1.service.TaskService;
+import com.itbangmodkradankanbanapi.service.TaskService;
 import jakarta.validation.Valid;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,9 +12,11 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+
 
 @RestController
-@RequestMapping("/v1/tasks")
+@RequestMapping("/v2/tasks")
 public class TaskController {
     @Autowired
     TaskService service;
@@ -23,14 +26,9 @@ public class TaskController {
     @Autowired
     private ModelMapper mapper;
 
-    @GetMapping("")
-    public ResponseEntity<Object> getAllTask() {
-        return ResponseEntity.ok(listMapper.mapList(service.findAllTask(), TaskDTO.class, mapper));
-    }
-
-    @GetMapping("/test")
-    public ResponseEntity<Object> getAllTask1() {
-        return ResponseEntity.ok(service.findAllTask());
+   @GetMapping("")
+    public ResponseEntity<Object> getAllTask(@RequestParam(required = false) List<String> filterStatuses , @RequestParam(required = false) String sortBy) {
+        return ResponseEntity.ok(listMapper.mapList(service.findAllTask(filterStatuses,sortBy), TaskDTO.class, mapper));
     }
 
     @GetMapping("/{id}")
@@ -39,7 +37,7 @@ public class TaskController {
     }
 
     @PostMapping("")
-    public  ResponseEntity<TaskDTO> createNewTask(@Valid @RequestBody TaskDTO task){
+    public  ResponseEntity<TaskDTO> createNewTask(@Valid @RequestBody TaskDTOForAdd task){
         TaskDTO createdTask = service.createNewTask(task);
         return ResponseEntity.status(HttpStatus.CREATED).body(createdTask);
     }
@@ -51,8 +49,8 @@ public class TaskController {
     }
 
     @PutMapping("/{id}")
-    public  ResponseEntity<TaskDTO> updateTask(@PathVariable Integer id ,@RequestBody TaskDTO taskDTO){
-        TaskDTO updatedTaskDTO = service.updateTask(id,taskDTO);
+    public  ResponseEntity<TaskDTO> updateTask(@PathVariable Integer id , @RequestBody TaskDTO taskDTO){
+        TaskDTO updatedTaskDTO = service.updateTask(id, taskDTO);
         return ResponseEntity.ok().body(updatedTaskDTO);
     }
     }
